+++ packages/ide-extension/src/decorations/propertiesMissingPreview.ts
@@ -2,7 +2,7 @@ import { state } from "../state.js"
 import * as vscode from "vscode"
 
 export const propertiesMissingPreview = () => {
-	const ideExtension = state().inlang.customApi()["app.inlang.ideExtension"]
+	const ideExtension = state().project.customApi()["app.inlang.ideExtension"]
 
 	const activeTextEditor = vscode.window.activeTextEditor
 	if (!activeTextEditor) {