+++ packages/versioned-interfaces/project-config/src/index.ts
@@ -1,1 +1,0 @@
-export { ProjectConfig } from "./interface.js"