+++ packages/versioned-interfaces/project-settings/src/index.ts
@@ -1,0 +1,1 @@
+export { ProjectSettings } from "./interface.js"